minion: Replace swarm with our own scheduler
All in all, Docker Swarm is a reasonable container orchestrator.
That said, it doesn't fit cleanly into the Quilt model, requiring some
pretty hack-ish gymnastics to get basic features like region based
placement.  Furthermore, it won't work in future when we'd like to do
placement based on availability, price, etc.

Given the current architecture of Quilt -- per node controllers, and
etcd for consensus -- implementing our own custom scheduler is fairly
simple.  This patch goes ahead and does it.

Note that we keep swarm running on the system as a convenient
debugging too.  It allows users to easily interact with all of the
containers running in the system. However, it isn't making any
placement decisions, instead it's operating in a read-only mode.