ovsdb: Redesign ovsdb with simpler interface and improved performance
Before this patch, there were many redundant ovsdb functions that
causes large amounts of unnecessary ovsdb transactions. That
becomes the performance bottleneck for quilt when booting a large
cluster. This patch simplifies the ovsdb interface, and reduces
O(N) number of ovsdb transactions in terms of number of VMs to O(1).

Co-Authored-By: Ethan J. Jackson <ejj@eecs.berkeley.edu>