minion/etcd: Rewrite network.go to make fewer etcd writes
Previously minion/etcd/network.go was inefficient and hard to
understand. Etcd is only capable of about 75 writes a second
with a single client and the old code could make a lot
of writes every time a container synced between the DB and
etcd. This patch changes that, with syncs making at most 3
writes to etcd in the worst case and 0 in the best case.
In addition, the code has been restructured to be easier to
read and modify in the future.