specs: Make specs' "CHANGE ME" fields consistent
Many of the specs have fields that need to be filled in by
the quilt user. These are rather inconsistent from spec to spec;
some have YOUR_GITHUB_USERNAME, some have <YOUR_GITHUB_USERNAME>
and some have the username of a quilt contributor. This patch
changes all of the githubKey fields to YOUR_GITHUB_USERNAME for
consistency and easy word selection when editing. It also makes
similar edits for some other fields.