docker: Pull name check logic out of Run()
Before this patch, Run would check if a container with the given name
was already running and skip booting it.  This logic is somewhat
specific to the supervisor, so it doesn't entirely make sense for it
to live in the docker client.

This patch does add some duplicate logic to supervisor.go, but it's in
a function that the scheduler is deleting entirely.