minion: Fold minion binary into the main quilt binary
In an effort to make things easier to keep track of, we've decided to
have a single quilt binary rather than splitting functionality across
a bunch of separate binaries.  For now, this means that the minion is
folded into the main quilt binary, in future new utilities will be
added there as well.