quilt-tester: Disable check_docker test
Go really doesn't like having multiple binaries live in a single
directory.  Since we're doing this with quilt-tester, this causes us
to have to do some rather awkward things to get going.  Fortunately,
the only test where we really need to do this at the moment is the
basic tests.  For now, we'll just disable one of the tests so it only
has one binary in it's directory.  Soon we'll refactor the
quilt-tester so this is no longer a problem.