quilt-tester: Gather command output before `Wait`ing
We were seeing intermittent test failures where not all of the stdout
and stderr output was gathered before returning. This was because
"wait will close the pipe after seeing the command exit". Thus, `Wait()`
would sometimes close the output pipes before the functions grabbing the
output had a chance to read the entire buffer.