docker: Re-pull images once every minute
When developing specs, it's quite helpful if we can update the
container images without having to completely re-create the cluster.
This patch makes this possible by invalidating the docker pull cache
once per minute.  Thus, to get all new containers, it should be
sufficient to just deploy an empty spec to remove the old containers,
and then deploy a new spec with the containers loaded.