quiltctl: Remove DaemonResponseError type
Before, we had an error type to represent when the daemon responded with
an error. This reduced the amount of repeated text when printing
errors. However, the error messages ended up being too general, and it
was easy to forget to wrap errors in this DaemonResponseError type.

We now log the error explicitly so that we can provide added context.