quiltctl: Automatically add host key when SSHing
This patch automatically adds an SSH option to disable checking whether
the target machine's fingerprint should be added to the user's
known_hosts. This is fairly reasonable because users wouldn't be able to
compare the fingerprint with anything else anyways.

Without this, users would have to acknowledge that the fingerprint is correct
everytime before SSHing into a new machine.