cluster: Run the foreman before cluster sync
When the user installs a new spec, it's ideal if that can be reflected
in the cluster as soon as possible.  By running the foreman before the
cluster sync, we can ship the new spec to already running machines
without having to wait for the rather expensive and heavy-weight
cluster.sync() function to complete.