network: Automatically create hostnames for containers
Before, containers within a label weren't directly addressable because only the
label as a whole had a hostname. However, applications often have to directly
talk to their peers (e.g., Spark workers have to talk to each Spark master
individually). This forced us to write ugly specs where a label is created for
each container just so that we could directly address them.

This patch makes it so that containers within a label automatically get the
hostname`<i>.<label>.q`.