network: Use AddressSets for enforcing ACLs
Before, we enforced ACLs by enumerating all the IPs that could
communicate. This led to very long, and redundant match strings. This
patch creates an AddressSet for each label, and then creates ACLs simply in
terms of the label name.

This significantly reduces the memory used for creating ACLs. Before, we
would run out of memory on the Master when booting 10,000 containers on
16 `m4.4xlarge`s.