vagrant: Log STDERR on failed shell command
This logs STDERR when a Vagrant shell command that was
intended to succeed (all of them currently) errors
(non-zero exit code). The Vagrant provider is currently
premature and faces a variety of errors -- all resulting in
"... is unavailable" or "... is unable to boot", but that
does not provide the user (or developer) much context to
debug the issue. This should help a lot.