vagrant: Revise error message when vagrantApi#Up fails
When `vagrant up` failed before, the message said we were
unable to check the machine status, but in reality, the real
problem is that we're unable to start the machine so we should
be clear about that. Status is a problem that the caller should
(and does) indicate (e.g. "provider %s is unavailable").