server: Return error if too many workers are specified
Now that the workers get a subnet in order to assign IPs, we have
a limit on the number of workers that can be created. Instead of
having one or more workers in a large deployment stall forever
attempting to grab a subnet, we error when trying to install a
stitch that would cause this scenario.