quiltctl: Refactor command parsing
Before, every command built its own flag set to be parsed. This meant
that common flags such as `host` had to be repeated for each command.
This patch refactors flag creation such that each command takes a flag
set, and sets the variables it cares about. This allows us to compose
the flag set creation.