etcd: Replace Update with Refresh
Previously, we were using the Update function to refresh the TTLs on
Etcd entries that had a limited lifetime. But, the Update function
would overwrite another entry with the same key, which is incorrect.
It has been replaced with the Refresh function, which only refreshes
the entry if it already exists and already has the requested value.

This patch also adds a similar function RefreshDir, which is needed
now that we support expiring directories for Etcd.