stitch: Rewrite language on top of Javascript
Javascript is easier to read, and more familiar to users than Lisp.
Embedding within an existing Javascript implementation also removes the
development overhead of writing language features not related to our
project, such as functions and mathematical operators.

The language is now represented as Javascript objects (e.g. Machine,
Container, Connection) inside a deployment object. There's one global deployment
object. To get the output of the spec, we evaluate the Javascript, and then
call the `toQuiltRepresentation` method to convert the deployment object
into a Javascript object that can be directly unmarshalled into the Go
structs. As before, The QRI then enforces the spec according to the
values in the context.

The semantics of the language have also slightly changed:
- There's an explicit deployment object.
    - Namespace, AdminACL, and MaxPrice are attributes of this object,
      rather than magic variables.
    - Machines and Labels must be explicitly deployed into this object.
- Label names are guaranteed to be unique by the language.
- It's much more object-oriented: e.g. to connect `foo` and `bar` on port 80,
  one would do `foo.connect(80, bar)` instead of `connect(80, foo, bar)`.

Co-Authored-by: Vivian Fang <vivian98.fang@gmail.com>