minion: Generate IPs less randomly
Before, we generated random IPs by picking a random number and
checking if the corresponding IP address was taken. This makes it
very hard to generate the last few available IP addresses. Instead,
we now pick a random IP to start at, and search linearly within the
given mask from there. This still maintains randomness, but is more
efficient.