providers: Rip out Azure for now
Overall, Azure seems to be a robust scalable cloud provider that's
quite popular, especially with enterprises.  That said, the Go API is
not mature, and user management on Azure is extremely painful.  It
seems that to get anything done requires operating across two separate
portals, each with inconsistent subsets of the required functionality.
Figuring out how to add a new user is an unbelievably difficult chore,
compounded by inconsistent (or buggy) behavior in the UI.  While
maintaining the existing users is no less difficult.

Given that currently the project isn't focused on enterprises, Azure's
main users, the benefit of maintaining the Azure provider seems far
outweighed by the developer time spent wrestling it into submission.
That's why, for now, we're removing support for it.  In future, when
the project has more developers, or when Azure's management interface
has improved, we will reconsider.