scheduler: Make exclusion constraints bidirectional
Before, when enforcing exclusion placement constraints, they were
only enforced in one direction. That is, if you had a placement
constraint that had TargetLabel = red and OtherLabel = blue, then
blue containers could be placed on a minion with red containers.
Visually, the constraint was Red -> Blue.

The correct behavior is for this constraint to work both ways,
so that blue containers could not be placed on a minion with red
containers. This patch corrects that behavior.