minion: Don't allow parallel boot and network config.
Docker is hanging rather frequently in the integration tests, and our
current theory is that the network configuration is occurring on
partially booted containers causing docker to behave badly.  As a
temporary fix (until the proper libnetwork plugin appears) we put the
docker boots and terminations into a transaction in an effort to
prevent the network code from running concurrently.  This should be
removed in the nearish future.