quiltctl: Don't add to `known_hosts` when SSHing
Before, we automatically added Quilt machines to `~/.ssh/known_hosts`
when SSHing into them with `quilt ssh`. This wasn't very useful because
we bypassed the prompt for users to confirm adding the host. Furthermore,
this polluted the `known_hosts` file. This resulted in eventual
collisions when cloud providers reused IPs.