specs: MEAN spec should wait for Mongo to boot
This adds mgo back as a dependency, but using govendor, it removes
the "internal/sasl" package from vendor/. This package is only
necessary if you build with `-tags sasl`, but godep likes to use
`./...` paths so it was picking it up anyways. Unfortunately,
that package requires cgo/gcc since it has native, non-Go dependencies.