stitch: Make QRI language-agnostic
Before, the QRI passed around and recompiled the full Javascript
specification to determine attributes such as how many machines and
containers to deploy. This patch modifies the QRI to just take a JSON
object that describes the deployment.

This formally separates the language from the deployment engine. It also
removes the need to compile imports into a single file in order to be
sent to minions.