engine: Set the default namespace from the Stitch
Setting the default namespace in the engine went against our general
principle of keeping complexity in the language.

This happens to fix a bug where we would error that we were changing
namespaces when using the default namespace. This was caused because we
would modify empty namespaces to be "default-namespace" before saving it to
the cluster. Thus, when another run request for the empty namespace came
in, we would think the namespace changed.

This also changes the `quilt stop` flag precedence to use the positional
argument over the named argument.