docker: Stop using docker exec
We're currently dealing with a docker bug in which the docker daemon
deadlocks occasionally.  While looking through stack traces we've
noticed that the docker daemon appears to be processing exec calls.
Additionally, a month or two ago we noticed that the minion was making
an exec call at the time of the deadlock.  This patch deals with the
issue by removing all exec calls from the minion.

A happy side effect of the approach this patch takes, is slightly
cleaner code.  All of the exec calls before this patch were to access
OVS utilities that are now bundled with the minion container directly.