travis: Build and push quilt/quilt and quilt/tester using travis
The docker hub build system isn't ideal for a lot of reasons.  Most
seriously, it's easy to end up in a situation where travis passes but
our containers fail, causing problems.  This patch rolls the docker
container build and push into the travis build, which should make
things a little bit more stable.  It also allows us to simplify the
Dockerfiles a lot.