scheduler: Safely unassign containers
Before, the scheduler deleted containers from a minion as the scheduler looped
over these containers. This could cause incorrect placements and going out of
bounds. With this patch, we instead keep a separate list of validly placed
containers, and assign it to the minion when all placements have been checked.