ovs: Upgrade to OVS 2.6 and use the OVS provided kernel module
OVS 2.6 is the first release of Open vSwitch with official support for
OVN.  The previous release, had more of a development preview with now
guarantees of stability or correctness.  In order to get OVS 2.6 to
work, we need several features from the kernel that aren't available
in stock Ubuntu 16.04.  This patch solves the problem by packaging the
official OVS 2.6 kernel module with the quilt/ovs container.  When our
VMs boot, we install the kernel module and get the requisite feature
(plus some other nice ones like STT).

Co-Authored-By: Kevin Lin <kaikai526@gmail.com>