network: Fix OpenFlow flip-flopping
Before, the flows responsible for forwarding packets from LOCAL to
containers were being created with a format that was changed by
`ovs-ofctl`. Specifically, actions of the form `actions=1` were being
changed to `actions=output:1`. This meant that on every run, we would
delete the `actions=output:1` flow, and try to replace it with
`actions=1`.

This patch fixes this bug by generating the target flows in a format
that won't be modified by `ovs-ofctl`.