scheduler: Pull container boots outside the transaction
We had temporarily moved container boots into the scheduler database
transaction as a hack to reduce the parallelism between booting and
networking containers.  This was in an effort to fix an occasional
docker daemon bug that was since fixed by commit
0907977 (docker: Stop using docker exec).  Now that the fix has been
merged, we ca start slowly and carefully re-enabling the parallelism.

This patch should also fix a problem in which downloading large
containers freezes the minion long enough for it to be removed from
etcd.