cluster: Refactor cluster package structure
This patch refactors the cluster, paying specific attention to the
packages and how they relate to each other.  In general, this patch
aims to the minimal necessary changes to the actual application
logic, instead just focusing on the interfaces between modules.

The new structure is simpler, easier to reason about, and provides a
separate namespace for each cloud provider (rather than forcing them
to rely on object methods for separating).  In order to avoid a rather
nasty circular dependency problem, the patch introduces several rather
small sub-packages which, though not ideal, are still a great
improvement over what was there.