cluster: Refactor cluster runOnce loop
This patch makes a number of simplifications and improvements to the
cluster runOnce loop designed to make the code more readable and
robust.  First, it combines the Machine and ACL reads into a single
transaction, guaranteeing that they apply to the same namespace.
Second, at the beginning of that transaction it verifies that the
cluster namespace hasn't changed, thus guaranteeing that we're
operating on the appropriate cluster.