minion: Configure container veths with a custom network driver
Currently, running an application with quilt required a
user to write special code in their application to wait for the
container it was running on to have its networking interfaces
configured. This is not ideal, so this patch runs containers using
a custom quilt network driver. Now, when booting a container, docker
will block until the driver configures the endpoints (veths) for it.

This does not completely eliminate the need for the loop that waits
for the network to come up, but it is the start of a series of
changes that will remove them.

A consequence of this is that it reduces the complexity of our
current networking code, which no longer needs to worry about veths
at all, since they are created and destroyed on demand, per each
container.