foreman: On Init, don't overwrite DB role with None
When the foreman first runs on a new cluster, it will overwrite the
database with the roles that the minions say they've been assigned in
previous runs.  This is to inform the engine what the actual roles
assignments are, so that it can choose to keep them stable if they're
compliant with the policy.  However, if a machine was booted
previously, but never had a role assigned to it, the foreman would
still overwrite the engine assigned role with None.  In this case, it
would be better to stick with whatever the engine had assigned us.