engine: Run whenever the database changes
The engine was designed to run periodically so that it may react to
changes in the local IP and, more importantly, changes to the machine
table based on new information from the foreman.  This patch changes
the engine code to use a more traditional database run loop.
Furthermore, it changes the API server to deploy specifications by
writing them to the cluster table from which the engine reads.