network: Lean on docker to set container routes
Our container routing requirements are fairly simply: the container
subnet should be 10.0.0.0/8, the IP should be allocated in the
minion's /20 within that /8, and the default route should be to the
network gateway, 10.0.0.1.  By properly configuring the docker network
driver, we can achieve all of these requirements without needing any
manual network route synchronization.

This patch also fixes a rather mysterious bug in which our route code
would occasionally designate the primary IP as a container's label IP
instead of the container's main IP.   This would cause networking to
completely fail for that container.  We don't fully understand what
was causing the issue, but the bug definitely resided in the code
deleted by this patch.