scheduler: Don't attempt to allocate container IPs
Rather misleadingly, when using the default IPAM driver, docker
ignores the container IP address suggested to it in the run command.
Instead, docker was allocating its own IP address for each container,
which the scheduler was dutifully adopting.  Since the scheduler IP
allocation code is essentially a noop, this patch removes it.