minion: Use IP to identify the default Gateway
Before this patch, were were identifying packets destined for the
default Gateway based on its MAC address.  This required us to query
the MAC address of the quilt bridge port so that we could set up the
OpenFlow rules to forward based on it.  This patch, instead, has us
identify packets destined for the default gateway based on its IP
address, 10.0.0.1.  That way, we can avoid having to look up the MAC
address, and can delete the associated logic.