network: Migrate `updateContainerIPs` to netlink
The `updateContainerIPs` function previously shelled out to the `ip`
utility to update container IPs.  Not only is this inefficient, it
also required a rather massive amount of ugly machinery to work.  This
patch instead calls netlink directly to set the container addresses
and, in the process, removes the `updateNamespaces` function which
becomes unnecessary.