haproxy: Simplify haproxy configuration
When I tested it locally, HAProxy was failing with a 503 error.  I
didn't really understand the config file, so I replaced it with
something much simpler taken from the HAProxy documentation.  This
happened to fix the error I was running into, so I figured we may as
well make it the default.