network: Remove concurrency from updateContainerIPs
The updateContainerIPs function was doing fork-join concurrency
splitting containers up across multiple go routines.  In future
patches we migrate to calling netlink directly instead of
through `ip`.  At this time, we'll have to lock the netlink calling
go routine to a particular OS thread, making the fork join parallelism
significantly more complicated to reason about.  Furthermore, the
benefits of the parallelism will likely decrease significantly, as we
no longer have the overhead of forking the `ip` process to split
across multiple cores.  For sake of simplicity, this patch removes the
parallelism for now.