network: Replace `network.VethPairNames` with `ipdef.IFName`
Since we've moved to the libnetwork plugin, the precise name of the
container interface has become unimportant.  As a result, the
`network.VethPairNames` function, which was always a bit ugly, seems
even more odd.

This patch replaces it with `ipdef.IFName` which has a much simpler
design.  Given a string, it simply truncates it so that it may be used
as an interface name.  Furthermore, it does this in a safer way by
checking the actual maxim interface size defined by the system in
`syscall.IFNAMSIZ` instead of hard coding a guess.