quiltctl: Attach stdin to SSH only if pty is used
Before this patch, a call to NativeClient.Run() attached os.Stdin
to the SSH session regardless of whether there was a psuedo-tty to
receive input.

When Stdin was attached without being used, a subsequent call to
NativeClient.Run() that actually read from Stdin would lose the
first character of keyboard input.