quiltctl: Make `machines` and `containers` similar
This patch makes various changes to the output format of `machines`
and `containers` so that they are suitable for being outputted one
after another in a future patch for `quilt ps`.

Before this patch, `quilt machines` outputted the String() value of
each Machine stored in DB, unlike `quilt containers` which outputted
a table using tabwriter.

After this patch:
   * `quilt machines` uses tabwriter.
   * The first column of `quilt containers` is changed from ID to
CONTAINER to be consistent with `quilt machines`.
   * A column called STITCH ID is placed inserted before the first
column of `quilt containers`.
   * The newline between the first and second row of `quilt
containers` is removed.