quiltctl: Various `quilt ps` tweaks
This patch adds PublicIP fields to `machines` and `containers`.

It also changes the first column of both `machines` and
`containers` from MACHINE or CONTAINER to ID.

In `containers` output, a blank line is now printed in between
containers hosted on different machines.

In `containers` output, the IMAGE and COMMAND fields are combined
into a single field, CONTAINER.

In `ps` output, "MACHINES" and "CONTAINERS" are printed before
the output of `machines` and `containers`.