mean: Timeout connection attempts to Mongo
Before, we blocked forever until the connection to Mongo succeeded. This
sometimes caused the MEAN app to hang forever because the connection attempt
started before the network was fully configured.

We now dial with a timeout. If the connection fails, we retry as before --
eventually, the network will be configured and the connection will succeed.