mongo: Refactor the Mongo DB run script
The mongo db run script had several problems which this patch fixes.
First, it waited for the network to come up before starting `mongod`.
This was no longer necessary, and failed occasionally due to a
peculiarity in the go resolver code.  Second, it misconfigured the
Mongo DB to bind to the IPs of remote instances, and finally third, it
was much more complex than necessary.