lint: Don't fail until after all files have been checked
In order to fail the build on lint errors, we have been exiting
the lint target if `golint` returns a non-zero exit code. This has
the annoying side effect of cancelling the scan when the first lint
error is found. This patch fixes that, so `make lint` will report
all lint errors before exiting with an appropriate code.