amazon: Change wait to require machines to be instantiated
When listing machines on Amazon, it was possible to get a response
after the assignment of cloud IDs but before machine initialization.
This caused the cluster's join to fail, booting twice as many
machines as needed.  The Amazon wait() function now requires machines
to have both a cloud ID and a size before returning.