db: Annotate db.Container for JSON
A future patch will write db.Container structures directly to etcd.
While not strictly necessary for correctness, it's nice to clean up
the JSON representation a bit by annotating the struct with some
hints.  This patch omits the ID field, and allows all other fields to
be omitted when empty.