minion: Reorganize IP management
Before this patch, IP addresses were allocated by the workers, who
then told the master about them so logical ports could be created.
This round trip needlessly complicated the code, and increased the
amount of time a container needed to wait for its network to be ready.

This patch, instead, allocates container IP address in the master
before communicating them to the minions. This has a number of
significant benefits:

    * DNS entries of every container in the cluster can be populated
      even before containers boot.

    * The master can create logical ports in parallel as the workers
      boot containers.

    * Container IPs need not be written to Etcd by the workers.

    * Minions no longer need allocate subnets. Both simplifying the
      minion etcd synchronization code and ipdef.

Also, in the process, this does a general refactoring of the code to
make it easier to reason about.