amazon: Fix multi-halt bug
When fixing a bug that caused machines to be booted multiple times,
another bug was introduced that caused them to be halted multiple
times. This was due to the new filter put in place in the wait()
function, as it was omitting machines as soon as their size
attribute was stripped. But, the cluster still saw the machines
when it used provider.List(), and so requested another halt of
the machines. This patch adds a simple condition to only look for
the size attribute when booting.