network: Properly forward broadcasts from the gateway
The previous openflow logic failed to forward broadcasts from the
gateway because register 1 (the associated veth) wasn't defined. To
remedy this, this patch adds a new rule that outputs to all veths when a
broadcast comes from the gateway.

This fixes a bug where public-facing containers weren't accessible
before the private network was exercised. This happened because ARP
packets coming from the gateway to resolve the IPs in the NAT iptable
rules were being dropped.

As a side-effect, this also fixes an old bug that caused public
connections to fail when two containers listened on the same machine.