minion: Record container creation time, status on workers
This patch allows for workers to record the time containers were started
by querying the docker client API. Workers can also record the status of
a container.

A time.Time field was added to db.Container. As it cannot be omitempty,
several tests were corrected for proper JSON output.

Various tests were updated as well to reflect the changes.