google: Fix error caused by subsequent inits
If a user who doesn't have GCE setup locally init'd the cluster
twice, the error wouldn't be returned by the client init function
the second time since we run the block that causes an error in
once.Do. This removes the clever code in favor of the simpler,
correct solution.