stitch: Assign machines a StitchID
Like containers, machines now have a StitchID based on the hash of the
machine attributes. This will allow us to make the UX for interacting
with machines and containers consistent. Additionally, the consistency
of the hash will allow us to consistently associate machines in the
cloud with those in the stitch.

Note that the join behavior is slightly changed by this patch.
Associating a StitchID with a consistent machine takes precedence over
some of the previous "fuzzy" matching for edge cases, such as whether a
machine has an IP.