ovsdb: Implement new helper function OpenFlowPorts()
The ovsdb.ListInterfaces() function is rather complex and
heavy-weight.  In a future patch, instead of doing a full join, it
will be sufficient to simply lookup the OpenFlow port number
associated with each interface.  This patch implements a new function
that does precisely that, so that ListInterfaces() may be deleted in a
future patch.