cluster: Add role to machine join
Since role is now determined at boot time, role must be added to the
join in cluster so that machines with different roles can be properly
torn down or booted.

As the CloudProviders do not know the role of each machine, the foreman
is queried to update the list of machines returned from the CloudProviders
with their roles.

Foreman no longer sends role in protobuf minion config. Foreman tests that
were testing this ability to change roles were removed.

Minion was also modified to not accept role from the foreman.