quiltctl: Support retrieving minion logs with `quiltctl logs`
Before, to retrive minion logs, users had to `quilt ssh` into the
minion machine and run `docker logs minion`. `quilt logs` now
determines if the target is a machine or a container, and retrieves
the minion logs if it's a machine, and the container logs otherwise.

We also updated the `quilt logs` tests to test each individual flag,
as opposed to testing them all together as was done before.