stitch: Fix bug with multiple identical services
Previously, when multiple identical services were deployed, you would
end up with inconsistent StitchID assignment causing bugs.

The bug can be reproduced by running a spec with a single spark cluster,
then running a spec with an additional identitcal spark cluster. With
the bug, the master container for the second cluster will be assigned
the original StitchID of the first cluster's master, causing all the new
workers to connect to the first cluster's master.

The bug arose from the `_refID` fields of containers being strings,
causing a container with a `_refID` of `"10"` to be assigned a StitchID
before a container with a `_refID` of `"8"`. We fix this bug by making
`_refID` a number so that StitchIDs are assigned in the correct order.