network: Store DNS mappings in new Hostname table
This patch modifies the DNS server to reference the new Hostname table
rather than the Label table. This separates the generation of hostnames
from serving them. This proves useful in a future patch that allows
assigning containers hostnames because the Label table will not be the
sole table determining hostname mappings.