stitch: Support setting container hostnames
Some specs require that containers know their own hostnames. For
example, etcd nodes have to publish their IP into the cluster.

Before this patch, there were two ways to achieve this. One was to
create a `Service` for each node; however, `Service`s are supposed to be
for load balancing. Another approach (that no longer works) was to
create a single `Service` with all the members, and infer the
container's IP address by mapping the result of `Service.children` onto
the hostnames of the containers. This approach was fragile because it
assumed determinism in the mapping of `Service.children` onto `Container`s
in the language, and the assignment of IP addresses in the QRI.

This patch introduces a simple solution. Users can give containers a
hostname using `Container.setHostname`, and can then assume that their
given hostname will route to that container.