minion: Support building Dockerfiles
This patch adds support for specifying raw Dockerfiles directly in a
spec, and building them in the Quilt cluster. This is useful for users
to avoid the overhead of building and pushing images to a third party
registry.

This feature is not meant for production use. Specifically, it does not
handle master failover well. Because all masters build their own
versions of images, if the leader goes down, the new leader will
direct the cluster to run its version of the image, causing all custom
containers to restart. Furthermore, the contents of the container could
potentially be different if the build process is non-deterministic.